Fix dependency by moving Dimname.{h,cpp} NamedTensor.{h,cpp} to core/
`ATen/core/Tensor.h` and `ATen/core/TensorMethods.h` both depend on
Dimname.h and NamedTensor.h. Therefore `Dimname.h` and `NamedTensor.h`
should really be in `ATen/core`. It's not a problem right now because
this dependency chain (core files cannot depend on non-core files) isn't
enforced in our OSS builds, but it is necessary to resolve this before
removing the BUILD_NAMEDTENSOR flag.

Test Plan
- [namedtensor ci]
