+++ aten/src/ATen/core/Dimname.cpp
@@ -1,5 +1,5 @@
 #ifdef BUILD_NAMEDTENSOR
-#include <ATen/Dimname.h>
+#include <ATen/core/Dimname.h>
 #include <c10/util/Exception.h>
 
 namespace at {
