+++ aten/src/ATen/core/NamedTensor.cpp
@@ -1,5 +1,5 @@
 #ifdef BUILD_NAMEDTENSOR
-#include <ATen/NamedTensor.h>
+#include <ATen/core/NamedTensor.h>
 #include <ATen/core/Tensor.h>
 #include <c10/util/C++17.h>
